+++ src/main/java/lotto/Lotto.java
@@ -1,12 +1,17 @@
 package lotto;
 
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 public class Lotto {
+    public static final int LOTTO_PRICE = 1000;
+    public static final long LOTTO_NUM_COUNT = 6;
 
     private final List<Integer> lottoNumbers;
 
     public Lotto(List<Integer> lottoNumbers) {
+        validateLottoNumbers(lottoNumbers);
         this.lottoNumbers = lottoNumbers;
     }
 
@@ -16,6 +21,17 @@ public class Lotto {
 
     public LottoMatch getLottoMatch(List<Integer> winningNumbers) {
         return LottoMatch.fromMatchCount((int) this.lottoNumbers.stream().filter(winningNumbers::contains).count());
+    }
+
+    private void validateLottoNumbers(List<Integer> numbers) {
+        if (numbers.size() != LOTTO_NUM_COUNT) {
+            throw new IllegalArgumentException("로또 번호는 6개여야 합니다.");
+        }
+
+        Set<Integer> uniqueNumbers = new HashSet<>(numbers);
+        if (uniqueNumbers.size() != LOTTO_NUM_COUNT) {
+            throw new IllegalArgumentException("로또 번호는 중복될 수 없습니다.");
+        }
     }
 
 }
