+++ src/main/java/view/LottoNumbersParser.java
@@ -1,4 +1,4 @@
-package lotto;
+package view;
 
 import java.util.Arrays;
 import java.util.List;
