+++ src/test/java/lotto/LottoTest.java
@@ -1,11 +1,13 @@
 package lotto;
 
 import org.junit.jupiter.api.Test;
+import view.LottoNumbersParser;
 
 import java.util.Arrays;
 import java.util.List;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
 public class LottoTest {
 
@@ -40,6 +42,20 @@ public class LottoTest {
         LottoMatch result = lotto.getLottoMatch(winningNumbers);
 
         assertThat(result).isEqualTo(LottoMatch.SIX);
+    }
+
+    @Test
+    void 중복되는_로또번호면_예외를_던진다() {
+        assertThatThrownBy(() -> {
+            new Lotto(Arrays.asList(1, 2, 3, 3, 4, 5));
+        }).isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @Test
+    void 로또번호가_6개초과되면_예외를_던진다() {
+        assertThatThrownBy(() -> {
+            new Lotto(Arrays.asList(1, 2, 3, 3, 4, 5, 5));
+        }).isInstanceOf(IllegalArgumentException.class);
     }
 
 }
